# speed up sm2 sign/verify