… crypto/ec/asm/ecp_sm2z256-x86_64.pl
… crypto/ec/ecp_sm2p256.c
… crypto/ec/ecp_sm2z256.c
… crypto/ec/ecp_sm2z256_table.c