Fix #11992

add precomputation for sm2, we can speed up sign and verify about 4 times

```
$ ./apps/openssl speed sm2
Doing 256 bits sign CurveSM2's for 10s: 88228 256 bits CurveSM2 signs in 9.95s 
Doing 256 bits verify CurveSM2's for 10s: 66358 256 bits CurveSM2 verify in 9.85s
version: 3.0.0-alpha3-dev
built on: built on: Fri May 29 03:10:52 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: cc -fPIC -arch x86_64 -O3 -Wall -DL_ENDIAN -DOPENSSL_PIC -D_REENTRANT -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffaf3bfffebffff:0x29c67af
                              sign    verify    sign/s verify/s
 256 bits SM2 (CurveSM2)   0.0001s   0.0001s   8867.1   6736.9
```