@mattcaswell 2020-06-01
Ping @InfoHunter!

I've not reviewed this, but noticed a couple of things:

1) I don't think we can claim that SM2 is a "nist" curve!!

2) You need to run "make update"